# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Billing Core API",
    "description": "Subscription billing system with payment processing",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auto-payments/test-subscription/{subscription_id}": {
      "post": {
        "tags": [
          "auto-payments"
        ],
        "summary": "Test Auto Payment For Subscription",
        "description": "\u041f\u0440\u043e\u0442\u0435\u0441\u0442\u0438\u0440\u043e\u0432\u0430\u0442\u044c \u0430\u0432\u0442\u043e\u043f\u043b\u0430\u0442\u0435\u0436 \u0434\u043b\u044f \u043a\u043e\u043d\u043a\u0440\u0435\u0442\u043d\u043e\u0439 \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438.\n\u041f\u043e\u043b\u0435\u0437\u043d\u043e \u0434\u043b\u044f \u043e\u0442\u043b\u0430\u0434\u043a\u0438.\n\nPOST /api/v1/auto-payments/test-subscription/{subscription_id}\n\nArgs:\n    subscription_id: ID \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438 \u0434\u043b\u044f \u0442\u0435\u0441\u0442\u0438\u0440\u043e\u0432\u0430\u043d\u0438\u044f\n\nReturns:\n    Dict \u0441 \u0440\u0435\u0437\u0443\u043b\u044c\u0442\u0430\u0442\u043e\u043c \u0442\u0435\u0441\u0442\u0438\u0440\u043e\u0432\u0430\u043d\u0438\u044f",
        "operationId": "test_auto_payment_for_subscription_api_v1_auto_payments_test_subscription__subscription_id__post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Test Auto Payment For Subscription Api V1 Auto Payments Test Subscription  Subscription Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auto-payments/simulate-subscription-ending/{subscription_id}": {
      "post": {
        "tags": [
          "auto-payments"
        ],
        "summary": "Simulate Subscription Ending",
        "description": "\u0421\u0438\u043c\u0443\u043b\u0438\u0440\u043e\u0432\u0430\u0442\u044c \u043e\u043a\u043e\u043d\u0447\u0430\u043d\u0438\u0435 \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438 (\u0443\u0441\u0442\u0430\u043d\u043e\u0432\u0438\u0442\u044c end_date \u043d\u0430 \u0441\u0435\u0433\u043e\u0434\u043d\u044f).\n\u041f\u043e\u043b\u0435\u0437\u043d\u043e \u0434\u043b\u044f \u0442\u0435\u0441\u0442\u0438\u0440\u043e\u0432\u0430\u043d\u0438\u044f \u0430\u0432\u0442\u043e\u043f\u043b\u0430\u0442\u0435\u0436\u0435\u0439 \u0431\u0435\u0437 \u043e\u0436\u0438\u0434\u0430\u043d\u0438\u044f \u0440\u0435\u0430\u043b\u044c\u043d\u043e\u0433\u043e \u043e\u043a\u043e\u043d\u0447\u0430\u043d\u0438\u044f.\n\nPOST /api/v1/auto-payments/simulate-subscription-ending/{subscription_id}\n\nArgs:\n    subscription_id: ID \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438\n\nReturns:\n    Dict \u0441 \u0440\u0435\u0437\u0443\u043b\u044c\u0442\u0430\u0442\u043e\u043c",
        "operationId": "simulate_subscription_ending_api_v1_auto_payments_simulate_subscription_ending__subscription_id__post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Simulate Subscription Ending Api V1 Auto Payments Simulate Subscription Ending  Subscription Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auto-payments/process-subscription/{subscription_id}": {
      "post": {
        "tags": [
          "auto-payments"
        ],
        "summary": "Process Single Subscription",
        "description": "\u0420\u0443\u0447\u043d\u043e\u0439 \u0437\u0430\u043f\u0443\u0441\u043a \u043e\u0431\u0440\u0430\u0431\u043e\u0442\u043a\u0438 \u043e\u0434\u043d\u043e\u0439 \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438 (\u043d\u043e\u0432\u0430\u044f \u0430\u0440\u0445\u0438\u0442\u0435\u043a\u0442\u0443\u0440\u0430).\n\u0417\u0430\u043f\u0443\u0441\u043a\u0430\u0435\u0442 Celery \u0437\u0430\u0434\u0430\u0447\u0443 process_single_subscription_payment.\n\nPOST /api/v1/auto-payments/process-subscription/{subscription_id}\n\nArgs:\n    subscription_id: ID \u043f\u043e\u0434\u043f\u0438\u0441\u043a\u0438 \u0434\u043b\u044f \u043e\u0431\u0440\u0430\u0431\u043e\u0442\u043a\u0438\n\nReturns:\n    Dict \u0441 task_id",
        "operationId": "process_single_subscription_api_v1_auto_payments_process_subscription__subscription_id__post",
        "parameters": [
          {
            "name": "subscription_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Subscription Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Process Single Subscription Api V1 Auto Payments Process Subscription  Subscription Id  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auto-payments/retry-payment/{payment_id}/{attempt}": {
      "post": {
        "tags": [
          "auto-payments"
        ],
        "summary": "Retry Auto Payment",
        "description": "\u0420\u0443\u0447\u043d\u043e\u0439 \u0437\u0430\u043f\u0443\u0441\u043a \u043f\u043e\u043f\u044b\u0442\u043a\u0438 \u0430\u0432\u0442\u043e\u0441\u043f\u0438\u0441\u0430\u043d\u0438\u044f (\u043d\u043e\u0432\u0430\u044f \u0430\u0440\u0445\u0438\u0442\u0435\u043a\u0442\u0443\u0440\u0430).\n\u0417\u0430\u043f\u0443\u0441\u043a\u0430\u0435\u0442 Celery \u0437\u0430\u0434\u0430\u0447\u0443 retry_auto_payment_attempt.\n\nPOST /api/v1/auto-payments/retry-payment/{payment_id}/{attempt}\n\nArgs:\n    payment_id: ID \u043f\u043b\u0430\u0442\u0435\u0436\u0430\n    attempt: \u041d\u043e\u043c\u0435\u0440 \u043f\u043e\u043f\u044b\u0442\u043a\u0438 (\u043d\u0430\u0447\u0438\u043d\u0430\u0435\u0442\u0441\u044f \u0441 1)\n\nReturns:\n    Dict \u0441 task_id",
        "operationId": "retry_auto_payment_api_v1_auto_payments_retry_payment__payment_id___attempt__post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Payment Id"
            }
          },
          {
            "name": "attempt",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Attempt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Retry Auto Payment Api V1 Auto Payments Retry Payment  Payment Id   Attempt  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}